Fixed linux build.
svn path=/trunk/libavg/; revision=2820

+++ src/player/OGLTexture.cpp
@@ -281,7 +281,7 @@ void OGLTexture::calcTexCoords()
 const int OGLTexture::getTexMemDim()
 {
     if (m_pf == YCbCr420p || m_pf == YCbCrJ420p) {
-        return m_TexSize.x * m_TexSize.y * 1.5;
+        return int(m_TexSize.x * m_TexSize.y * 1.5);
     }
     else {
         return m_TexSize.x * m_TexSize.y * Bitmap::getBytesPerPixel(m_pf);
